requirements.txt
tensorflow
tensorflow_hub
numpy
opencv-python
Pillow
ipython